data_warehouse_app/backend/requirements.txt
alembic==1.16.2
biopython==1.85
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
